+++ packages/lesswrong/components/ea-forum/users/EAUsersProfile.tsx
@@ -15,7 +15,7 @@ import { taglineSetting } from '../../common/HeadTags';
 import { getBrowserLocalStorage } from '../../async/localStorageHandlers';
 import { siteNameWithArticleSetting, taggingNameIsSet, taggingNameCapitalSetting } from '../../../lib/instanceSettings';
 import { DEFAULT_LOW_KARMA_THRESHOLD } from '../../../lib/collections/posts/views'
-import { SORT_ORDER_OPTIONS } from '../../../lib/collections/posts/schema';
+import { SORT_ORDER_OPTIONS } from '../../../lib/collections/posts/sortOrderOptions';
 import { CAREER_STAGES, PROGRAM_PARTICIPATION, SOCIAL_MEDIA_PROFILE_FIELDS } from '../../../lib/collections/users/custom_fields';
 import { socialMediaIconPaths } from '../../form-components/PrefixedInput';
 import { eaUsersProfileSectionStyles, UserProfileTabType } from './modules/EAUsersProfileTabbedSection';
