+++ packages/lesswrong/lib/collections/tags/schema.ts
@@ -8,7 +8,7 @@ import GraphQLJSON from 'graphql-type-json';
 import moment from 'moment';
 import { captureException } from '@sentry/core';
 import { forumTypeSetting, taggingNamePluralSetting, taggingNameSetting } from '../../instanceSettings';
-import { SORT_ORDER_OPTIONS, SettingsOption } from '../posts/schema';
+import { SORT_ORDER_OPTIONS, SettingsOption } from '../posts/sortOrderOptions';
 import omit from 'lodash/omit';
 
 const formGroups: Partial<Record<string,FormGroup>> = {
